fixing memory error: \'funk2_memorypool__load_from_stream error: decompressed memory image is not the correct size.\'

+++ c/f2_zlib.c
@@ -26,7 +26,7 @@
 
 // *dest_length returns length of src_data after compression.
 // dest_data can be NULL.
-boolean_t zlib__deflate(u8* dest_data, int* dest_length, u8* src_data, int src_length) {
+boolean_t zlib__deflate(u8* dest_data, u64* dest_length, u8* src_data, u64 src_length) {
   int      zlib_result;
   int      byte_num;
   z_stream zlib_stream;
@@ -68,7 +68,7 @@ boolean_t zlib__deflate(u8* dest_data, int* dest_length, u8* src_data, int src_l
   return boolean__false;
 }
 
-boolean_t zlib__deflate_length(u8* src_data, int src_length, int* dest_length) {
+boolean_t zlib__deflate_length(u8* src_data, u64 src_length, u64* dest_length) {
   return zlib__deflate(NULL, dest_length, src_data, src_length);
 }
 
@@ -77,11 +77,11 @@ f2ptr f2__string__deflate(f2ptr cause, f2ptr this) {
     return f2larva__new(cause, 1, nil);
   }
   
-  int src_length = raw__string__length(cause, this);
+  u64 src_length = raw__string__length(cause, this);
   u8* src_data   = (u8*)from_ptr(f2__malloc(src_length));
   raw__string__str_copy(cause, this, src_data);
   
-  int dest_length = 0;
+  u64 dest_length = 0;
   if (zlib__deflate_length(src_data, src_length, &dest_length)) {
     f2__free(to_ptr(src_data));
     return nil;
@@ -101,7 +101,7 @@ f2ptr f2__string__deflate(f2ptr cause, f2ptr this) {
 }
 def_pcfunk1(string__deflate, this, return f2__string__deflate(this_cause, this));
 
-boolean_t zlib__inflate(u8* dest_data, int* dest_length, u8* src_data, int src_length) {
+boolean_t zlib__inflate(u8* dest_data, u64* dest_length, u8* src_data, u64 src_length) {
   int           zlib_result;
   unsigned int  byte_num;
   z_stream      zlib_stream;
@@ -155,7 +155,7 @@ boolean_t zlib__inflate(u8* dest_data, int* dest_length, u8* src_data, int src_l
   }
 }
 
-boolean_t zlib__inflate_length(u8* src_data, int src_length, int* dest_length) {
+boolean_t zlib__inflate_length(u8* src_data, u64 src_length, u64* dest_length) {
   return zlib__inflate(NULL, dest_length, src_data, src_length);
 }
 
@@ -164,11 +164,11 @@ f2ptr f2__string__inflate(f2ptr cause, f2ptr this) {
     return f2larva__new(cause, 1, nil);
   }
   
-  int src_length = raw__string__length(cause, this);
+  u64 src_length = raw__string__length(cause, this);
   u8* src_data   = (u8*)from_ptr(f2__malloc(src_length));
   raw__string__str_copy(cause, this, src_data);
   
-  int dest_length = 0;
+  u64 dest_length = 0;
   if (zlib__inflate_length(src_data, src_length, &dest_length)) {
     f2__free(to_ptr(src_data));
     return nil;
